build: rpath linker flag for native build
Add rpath linker flag for installing libraries to a none standard path
When building and installing dependent libraries (from common/framework) to a non
standard installation path which are linked to by the agent,
Running the agent requires setting LD_LIBRARY_PATH to point to the
dependent libraries installation path.

This commit update the agent's CMakeLists.txt to always add
-rpath=${CMAKE_INSTALL_FULL_LIBDIR} to the CMAKE linker flags - no
effect for cross compilation (since CMAKE_INSTALL_FULL_LIBDIR is already
in the linker search path), and remove the need to set LD_LIBRARY_PATH
before running the agent when installing dependencies to a custom path.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>